+++ src/argus/incident/apps.py
@@ -1,5 +1,5 @@
 from django.apps import AppConfig
-from django.db.models.signals import post_delete, post_save, pre_save
+from django.db.models.signals import post_delete, post_save
 
 
 class IncidentConfig(AppConfig):
@@ -9,11 +9,9 @@ class IncidentConfig(AppConfig):
     def ready(self):
         from .signals import (
             close_token_incident,
-            create_change_events,
             create_first_event,
             delete_associated_user,
             delete_associated_event,
-            detect_changes,
             send_notification,
         )
 
@@ -23,5 +21,3 @@ class IncidentConfig(AppConfig):
         post_save.connect(close_token_incident, "authtoken.Token")
         post_save.connect(create_first_event, "argus_incident.Incident")
         post_save.connect(send_notification, "argus_incident.Event", dispatch_uid="send_notification")
-        pre_save.connect(detect_changes, "argus_incident.Incident"),
-        post_save.connect(create_change_events, "argus_incident.Incident"),