+++ src/argus/incident/ticket/base.py
@@ -5,8 +5,11 @@ from abc import ABC, abstractmethod
 from django.conf import settings
 
 __all__ = [
+    "TicketClientException",
+    "TicketCreationException",
     "TicketPlugin",
     "TicketPluginException",
+    "TicketSettingsException",
 ]
 
 
@@ -14,27 +17,47 @@ class TicketPluginException(Exception):
     """Ticket plugin exception"""
 
 
+class TicketSettingsException(TicketPluginException):
+    """Some of the settings required for the ticket plugins are not set
+    properly
+    """
+
+
+class TicketClientException(TicketPluginException):
+    """There was an error creating the client for interaction with the ticket
+    system
+    """
+
+
+class TicketCreationException(TicketPluginException):
+    """There was an error creating a ticket in the ticket system"""
+
+
 class TicketPlugin(ABC):
     @classmethod
     def import_settings(cls):
         """
         Imports, validates and returns settings needed for ticket creation.
-        Raises a ValueError if settings are not set properly
+
+        Raises a TicketSettingsException if settings are not set
+        properly.
         """
         endpoint = getattr(settings, "TICKET_ENDPOINT", None)
         if not endpoint:
-            raise ValueError("No endpoint can be found in the settings. Please update the setting 'TICKET_ENDPOINT'.")
+            raise TicketSettingsException(
+                "No endpoint can be found in the settings. Please check and update the     setting 'TICKET_ENDPOINT'."
+            )
 
         authentication = getattr(settings, "TICKET_AUTHENTICATION_SECRET", None)
         if not authentication:
-            raise ValueError(
-                "No authentication information can be found in the settings. Please update the setting 'TICKET_AUTHENTICATION_SECRET'."
+            raise TicketSettingsException(
+                "No correct authentication information can be found in the settings. Please check and update the setting 'TICKET_AUTHENTICATION_SECRET'."
             )
 
         ticket_information = getattr(settings, "TICKET_INFORMATION", None)
         if not ticket_information:
-            raise ValueError(
-                "No extra information for ticket creation can be found in the settings. Please update the setting 'TICKET_INFORMATION'."
+            raise TicketSettingsException(
+                "No correct extra information for ticket creation can be found in the settings. Please check and update the setting 'TICKET_INFORMATION'."
             )
 
         return endpoint, authentication, ticket_information
@@ -46,17 +69,17 @@ class TicketPlugin(ABC):
         Creates, authenticates and returns a client to connect with the given
         endpoint
 
-        Raises a TicketPluginException in case of any errors
+        Raises a TicketClientException in case of any errors
         """
         return None
 
     @classmethod
     @abstractmethod
-    def create_ticket(cls, incident):
+    def create_ticket(cls, serialized_incident: dict):
         """
         Creates a ticket with the information of the given incident and returns
         its url
 
-        Raises a TicketPluginException in case of any errors
+        Raises a TicketCreationException in case of any errors
         """
         return None