+++ CHANGELOG.md
@@ -4,6 +4,24 @@ read by developers.
 
 ## [Unreleased]
 
+## [1.10.2]
+
+### Changed
+- Renamed the ticket creation endpoint via plugin from `/ticket/` to `/automatic-ticket/`
+- Refactored the view tests for NotificationProfile
+- Explicitly set language in sphinx conf, silencing a warning
+
+## [1.10.1] - 2022-12-08
+
+### Fixed
+
+- Run API auth V1 tests
+
+### Changed
+
+- Send serialized incidents to the ticket-plugin, not database objects
+- Improved version-detection in git flow
+
 ## [1.10.0] - 2022-11-17
 
 ### Added